+++ EveShopping.Modelo/Models/eshFittingHardware.cs
@@ -10,7 +10,7 @@ namespace EveShopping.Modelo.Models
         public int typeID { get; set; }
         public int slotID { get; set; }
         public short positionInSlot { get; set; }
-        public short units { get; set; }
+        public int units { get; set; }
         public double volume { get; set; }
         public virtual eshFitting eshFitting { get; set; }
         public virtual eshFittingSlot eshFittingSlot { get; set; }
