+++ EveShopping.Modelo/Models/eshShoppingListInvType.cs
@@ -7,7 +7,7 @@ namespace EveShopping.Modelo.Models
     {
         public int shoppingListID { get; set; }
         public int typeID { get; set; }
-        public short units { get; set; }
+        public int units { get; set; }
         public double volume { get; set; }
         public virtual eshShoppingList eshShoppingList { get; set; }
         public virtual invType invType { get; set; }
